feat!: Simplified Response
We switch to the use of only one response pattern, irrespective of whether the current representation is sent or not. An empty body or one with headers only (akin to a response to) may be sent in the first body part.

This is both simpler to generate for the middleware and clients to consume.

E-Tag header field has been removed from the requirements as this is specific to the current representation of the resource, which might be sent in the first body parts, and not the full response body that also includes notifications.

Signed-off-by: Rahul Gupta <CxRes@users.noreply.github.com>